Testing2-->Main

+++ app/views/productionmanager/pmorders.php
@@ -36,17 +36,54 @@
         .blue {
             background-color: #3498db;
         }
+        .yellow{
+            background-color: #f1c40f;
+            margin-left: 10px;
+        }
+        
     </style>
     <title>Customer Orders</title>
 </head>
 <body>
-    <?php include 'pmnavbar.php'; ?>
+
+    <?php include 'pmnavbar.php'; 
+    
+    if (isset($error)) {
+        echo '<script>
+            document.addEventListener("DOMContentLoaded", function () {
+                const showAlert = async () => {
+                    const SwalwithButton = Swal.mixin({
+                        customClass: {
+                            confirmButton: "greenbutton",
+                        },
+                        buttonsStyling: false
+                    });
+    
+                    if (typeof Swal !== "undefined") {
+                        await SwalwithButton.fire({
+                            icon: "error",
+                            title: "Oops...",
+                            text: "' . $error . '",
+                            confirmButtonText: "OK",
+                        });
+                        window.location.href = "' . BASE_URL . 'pmcontrols/index";
+                    } else {
+                        alert("' . $error . '");
+                    }
+                };
+    
+                // Call the async function to show the alert
+                showAlert();
+            });
+        </script>';
+    }
+    ?>
 
     <div class="searchpanel">
         <form method="GET" action="<?php echo BASE_URL;?>OrderControls/searchOrders" class="search" style="display:flex; flex-direction:row;">
         <?php
         if(isset($_GET['search'])){
-            echo '<input type="text" id="search" name="search" placeholder="Enter Order ID or Place BY" value="'.$_GET['search'].'" class="searchbox">';
+            echo '<input type="text" id="search" name="search" placeholder="Enter Order Ref or Place BY" value="'.$_GET['search'].'" class="searchbox">';
             echo '<input type="submit" value="Search" class="button green">';
             echo '<button class="button green" onclick="clearSearch(); return false;">Clear Search</button>';
         }
@@ -83,8 +120,7 @@
                 <th>Payment Status</th>
                 <th>Order Status</th>
                 <th>Total</th>
-                <th>Deliver By</th>
-                <th>Unique ID</th>
+                <th>Order Capacity(No.of Containers)</th>
                 <th>Deliver Address</th>
                 <th>Update Order</th>
                 <th>Cancel Order</th>
@@ -92,7 +128,7 @@
             </tr>";
             
             foreach ($productorder as $ProductOrder){
-                if ($ProductOrder->orderstatus == "pending"){
+                if ($ProductOrder->orderstatus == "pending" && $ProductOrder->orderdate >= date('Y-m-d')){
                     echo "<tr>";
                     echo "<td>".$ProductOrder->orderref."</td>";
                     echo "<td>".$ProductOrder->placeby."</td>";
@@ -100,14 +136,14 @@
                     echo "<td>".$ProductOrder->paymentstatus."</td>";
                     echo "<td>".$ProductOrder->orderstatus."</td>";
                     echo "<td>".$ProductOrder->total."</td>";
-                    echo "<td>".$ProductOrder->deliverby."</td>";
-                    echo "<td>".$ProductOrder->unique_id."</td>";
+                    echo "<td>".$ProductOrder->order_cap."</td>";
                     echo "<td>".$ProductOrder->deliver_address."</td>";
-                    if($ProductOrder->paymentstatus == "paid" || $ProductOrder->paymentstatus == "advanced"){
+                    if($ProductOrder->paymentstatus === "paid" || $ProductOrder->paymentstatus === "advanced"){
                         echo "<td><button class='button green' onclick='process(".$ProductOrder->orderid.")'>Process</button></td>";
                     }
                     else{
-                        echo "<td>Incomplete Payment</td>";
+                        echo "<td><button class='button green' onclick='validateProcess()'>Process</button></td>";
+                    
                     }
                     echo "<td><button class='button red' onclick='cancel(".$ProductOrder->orderid.")'>Cancel</button></td>";
                     echo "<td><button class='button blue' onclick='more(\"" . $ProductOrder->unique_id . "\")'>More</button></td>";
@@ -130,8 +166,8 @@
                     <th>Payment Status</th>
                     <th>Order Status</th>
                     <th>Total</th>
-                    <th>Deliver By</th>
-                    <th>Unique ID</th>
+                    <th>Order Capacity
+                    (No.of Containers)</th>
                     <th>Deliver Address</th>
                     <th>Update Order</th>
                     <th>Cancel Order</th>
@@ -139,7 +175,7 @@
                 </tr>";
                 
                 foreach ($productorder as $ProductOrder){
-                    if ($ProductOrder->orderstatus == "processing" && ($ProductOrder->paymentstatus == "paid" || $ProductOrder->paymentstatus == "advanced")  && $ProductOrder->orderdate == date('Y-m-d', strtotime('+1 day'))){
+                    if ($ProductOrder->orderstatus == "processing" && ($ProductOrder->paymentstatus == "paid" || $ProductOrder->paymentstatus == "advanced") ){ //&& $ProductOrder->orderdate == date("Y-m-d", strtotime('+1 day')) 
                         echo "<tr>";
                         echo "<td>".$ProductOrder->orderref."</td>";
                         echo "<td>".$ProductOrder->placeby."</td>";
@@ -147,10 +183,9 @@
                         echo "<td>".$ProductOrder->paymentstatus."</td>";
                         echo "<td>".$ProductOrder->orderstatus."</td>";
                         echo "<td>".$ProductOrder->total."</td>";
-                        echo "<td>".$ProductOrder->deliverby."</td>";
-                        echo "<td>".$ProductOrder->unique_id."</td>";
+                        echo "<td>".$ProductOrder->order_cap."</td>";
                         echo "<td>".$ProductOrder->deliver_address."</td>";
-                        echo "<td><button class='button green' onclick='completeProduction(".$ProductOrder->orderid.", \"".$ProductOrder->deliverystatus."\")'>Process</button></td>";
+                        echo "<td><button class='button green' onclick='completeProduction(".$ProductOrder->orderid.", \"".$ProductOrder->deliverystatus."\")'>Production Complete</button></td>";
                         echo "<td><button class='button red' onclick='cancel(".$ProductOrder->orderid.")'>Cancel</button></td>";
                         echo "<td><button class='button blue' onclick='more(\"" . $ProductOrder->unique_id . "\")'>More</button></td>";
                         echo "</tr>";
@@ -172,7 +207,8 @@
                 <th>Payment Status</th>
                 <th>Order Status</th>
                 <th>Total</th>
-                <th>Unique ID</th>
+                <th>Order Capacity
+                    (No.of Containers)</th>
                 <th>Deliver Address</th>
                 <th>Complete Order</th>
                 <th>Cancel Order</th>
@@ -189,7 +225,7 @@
                     echo "<td>".$ProductOrder->paymentstatus."</td>";
                     echo "<td>".$ProductOrder->orderstatus."</td>";
                     echo "<td>".$ProductOrder->total."</td>";
-                    echo "<td>".$ProductOrder->unique_id."</td>";
+                    echo "<td>".$ProductOrder->order_cap."</td>";
                     echo "<td>".$ProductOrder->deliver_address."</td>";
                     echo "<td><button class='button green' onclick='completed(".$ProductOrder->orderid.")'>Complete</button></td>";
                     echo "<td><button class='button red' onclick='cancel(".$ProductOrder->orderid.")'>Cancel</button></td>";
@@ -212,8 +248,7 @@
                 <th>Payment Status</th>
                 <th>Order Status</th>
                 <th>Total</th>
-                <th>Deliver By</th>
-                <th>Unique ID</th>
+                <th>Order Capacity(No.of Containers)</th>
                 <th>Deliver Address</th>
                 <th>Assign Vehicle</th>
                 <th>Cancel Order</th>
@@ -222,7 +257,7 @@
 
             foreach($productorder as $ProductOrder){ 
 
-                if($ProductOrder->orderstatus == "finishedproduction" && ($ProductOrder->paymentstatus == "paid" || $ProductOrder->paymentstatus == "advanced") && ($ProductOrder->deliverystatus == "delivery") && $ProductOrder->orderdate== date('Y-m-d', strtotime('+1 day'))) { 
+                if($ProductOrder->orderstatus == "finishedproduction" && ($ProductOrder->paymentstatus == "paid" || $ProductOrder->paymentstatus == "advanced") && ($ProductOrder->deliverystatus == "delivery") ) { //&& $ProductOrder->orderdate == date('Y-m-d', strtotime('+1 day'))
 
                     echo "<tr>";
                     echo "<td>".$ProductOrder->orderref."</td>";
@@ -231,10 +266,9 @@
                     echo "<td>".$ProductOrder->paymentstatus."</td>";
                     echo "<td>".$ProductOrder->orderstatus."</td>";
                     echo "<td>".$ProductOrder->total."</td>";
-                    echo "<td>".$ProductOrder->deliverby."</td>";
-                    echo "<td>".$ProductOrder->unique_id."</td>";
+                    echo "<td>".$ProductOrder->order_cap."</td>";
                     echo "<td>".$ProductOrder->deliver_address."</td>";
-                    echo "<td><button class='button green' onclick='assignvehicle(".$ProductOrder->orderid.")'>Assign</button></td>"; //, \"".$ProductOrder->order_cap."\"
+                    echo "<td><button class='button green' onclick='assignvehicle(".$ProductOrder->orderid.")'>Assign</button></td>"; 
                     echo "<td><button class='button red' onclick='cancel(".$ProductOrder->orderid.")'>Cancel</button></td>";
                     echo "<td><button class='button blue' onclick='more(\"" . $ProductOrder->unique_id . "\")'>More</button></td>";
                     echo "</tr>";
@@ -256,7 +290,7 @@
                 <th>Order Status</th>
                 <th>Total</th>
                 <th>Deliver By</th>
-                <th>Unique ID</th>
+                <th>Order Capacity(No.of Containers)</th>
                 <th>Deliver Address</th>
                 <th>Complete Order</th>
                 <th>Cancel Order</th>
@@ -275,7 +309,7 @@
                 echo "<td>".$ProductOrder->orderstatus."</td>";
                 echo "<td>".$ProductOrder->total."</td>";
                 echo "<td>".$ProductOrder->deliverby."</td>";
-                echo "<td>".$ProductOrder->unique_id."</td>";
+                echo "<td>".$ProductOrder->order_cap."</td>";
                 echo "<td>".$ProductOrder->deliver_address."</td>";
                 echo "<td><button class='button green' onclick='completed(".$ProductOrder->orderid.")'>Complete</button></td>";
                 echo "<td><button class='button red' onclick='cancel(".$ProductOrder->orderref.")'>Cancel</button></td>";
@@ -299,7 +333,7 @@
                 <th>Order Status</th>
                 <th>Total</th>
                 <th>Deliver By</th>
-                <th>Unique ID</th>
+                <th>Order Capacity(No.of Containers)</th>
                 <th>Deliver Address</th>
                 <th>More Details</th>
             </tr>";
@@ -315,7 +349,7 @@
                 echo "<td>".$ProductOrder->orderstatus."</td>";
                 echo "<td>".$ProductOrder->total."</td>";
                 echo "<td>".$ProductOrder->deliverby."</td>";
-                echo "<td>".$ProductOrder->unique_id."</td>";
+                echo "<td>".$ProductOrder->order_cap."</td>";
                 echo "<td>".$ProductOrder->deliver_address."</td>";
                 echo "<td><button class='button blue' onclick='more(\"" . $ProductOrder->unique_id . "\")'>More</button></td>";
                 echo "</tr>";
@@ -337,6 +371,8 @@
                 <th>Order Status</th>
                 <th>Total</th>
                 <th>Deliver By</th>
+                <th>Order Capacity 
+                (No.of Containers)</th>
                 <th>Unique ID</th>
                 <th>Deliver Address</th>
                 <th>More Details</th>
@@ -353,6 +389,7 @@
                 echo "<td>".$ProductOrder->orderstatus."</td>";
                 echo "<td>".$ProductOrder->total."</td>";
                 echo "<td>".$ProductOrder->deliverby."</td>";
+                echo "<td>".$ProductOrder->order_cap."</td>";
                 echo "<td>".$ProductOrder->unique_id."</td>";
                 echo "<td>".$ProductOrder->deliver_address."</td>";
                 echo "<td><button class='button blue' onclick='more(\"" . $ProductOrder->unique_id . "\")'>More</button></td>";
@@ -403,9 +440,34 @@
     }
 
     function process(orderid){
-        var url = BASE_URL + "pmcontrols/processOrder/" + orderid;
-        window.location.href = url;
-    }
+        const SwalwithButton = Swal.mixin({
+            customClass: {
+                confirmButton: "button yellow",
+                cancelButton: "button yellow"
+            },
+            buttonsStyling: false
+            });
+
+            SwalwithButton.fire({
+                text: "Are you sure you want to process this order?",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Yes",
+                cancelButtonText: "No",
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Order Processed Successfully',
+                        text: 'The order has been successfully processed.',
+                        confirmButtonText: 'OK'
+                    }).then(() => {
+                    var url = BASE_URL + "pmcontrols/processOrder/" + orderid;
+                    window.location.href = url;
+                    });
+                }
+            });
+    } 
 
     function completeProduction(orderid, deliverystatus) {
         if(deliverystatus=="pickup"){
@@ -419,9 +481,35 @@
     }
 
     function cancel(orderid) {
-        var url = BASE_URL + "pmcontrols/cancelOrder/" + orderid;
-        window.location.href = url;
+        const SwalwithButton = Swal.mixin({
+            customClass: {
+                confirmButton: "button yellow",
+                cancelButton: "button yellow"
+            },
+            buttonsStyling: false
+            });
+
+            SwalwithButton.fire({
+                text: "Are you sure you want to cancel this order?",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Yes",
+                cancelButtonText: "No",
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Order Canceled Successfully',
+                        text: 'The order has been successfully canceled.',
+                        confirmButtonText: 'OK'
+                    }).then(() => {
+                        var url = BASE_URL + "pmcontrols/cancelOrder/" + orderid;
+                        window.location.href = url;
+                    });
+                }
+            });
     }
+
 
     function more(unique_id) {
         var url = BASE_URL + "OrderControls/moreDetails/" + unique_id;
@@ -435,9 +523,34 @@
     }
 
     function completed(orderid){
-        sessionStorage.setItem('activeLink', 'showCompletedOrderTable(this)')
-        var url = BASE_URL + "pmcontrols/completeOrder/" + orderid;
-        window.location.href = url;
+        const SwalwithButton = Swal.mixin({
+            customClass: {
+                confirmButton: "button yellow",
+                cancelButton: "button yellow"
+            },
+            buttonsStyling: false
+            });
+
+            SwalwithButton.fire({
+                text: "Are you sure you want to complete this order?",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Yes",
+                cancelButtonText: "No",
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Order Completed Successfully',
+                        text: 'The order has been successfully Completed.',
+                        confirmButtonText: 'OK'
+                    }).then(() => {
+                    sessionStorage.setItem('activeLink', 'showCompletedOrderTable(this)')
+                    var url = BASE_URL + "pmcontrols/completeOrder/" + orderid;
+                    window.location.href = url;
+                    });
+                }
+            });  
     }
 
     function showPendingOrderTable(link){
@@ -521,6 +634,14 @@
         window.location.href = BASE_URL + "pmcontrols/index";
     }
 
+    function validateProcess(){
+            Swal.fire({
+                icon: 'error',
+                title: 'Order Cannot Be Processed',
+                text: 'No advance or full payment received',
+                confirmButtonText: 'OK'
+            });
+        }
     </script>
 </body>
 </html>
